+++ src/main/java/net/yxiao233/ifeu/common/provider/ModSerializableProvider.java
@@ -24,6 +24,7 @@ import net.minecraft.world.level.material.Fluids;
 import net.minecraftforge.common.Tags;
 import net.minecraftforge.fluids.FluidStack;
 import net.yxiao233.ifeu.IndustrialForegoingExtraUpgrades;
+import net.yxiao233.ifeu.api.recipe.builder.IFEURecipeBuilders;
 import net.yxiao233.ifeu.api.recipe.builder.ShapedRecipeBuilder;
 import net.yxiao233.ifeu.common.recipe.ArcaneDragonEggForgingRecipe;
 import net.yxiao233.ifeu.common.recipe.ShapedRecipe;
@@ -44,8 +45,8 @@ public class ModSerializableProvider extends TitaniumSerializableProvider {
     @Override
     public void add(Map<IJsonFile, IJSONGenerator> map) {
         //dissolution chamber
-        new DissolutionChamberRecipe(IFEU("liquid_dragon_breath"),
-                new Ingredient.Value[]{
+        IFEURecipeBuilders.dissolutionChamberRecipe(new ItemStack(Items.GLASS_BOTTLE,8),"liquid_dragon_breath")
+                .inputs(
                         itemValue(Items.DRAGON_BREATH.getDefaultInstance()),
                         itemValue(Items.DRAGON_BREATH.getDefaultInstance()),
                         itemValue(Items.DRAGON_BREATH.getDefaultInstance()),
@@ -53,14 +54,15 @@ public class ModSerializableProvider extends TitaniumSerializableProvider {
                         itemValue(Items.DRAGON_BREATH.getDefaultInstance()),
                         itemValue(Items.DRAGON_BREATH.getDefaultInstance()),
                         itemValue(Items.DRAGON_BREATH.getDefaultInstance()),
-                        itemValue(Items.DRAGON_BREATH.getDefaultInstance()),
-                },
-                new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),500),200,
-                new ItemStack(Items.GLASS_BOTTLE,8),new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),100)
-        );
+                        itemValue(Items.DRAGON_BREATH.getDefaultInstance())
+                )
+                .inputFluid(new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),500))
+                .processingTime(200)
+                .outputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),100))
+                .save();
 
-        new DissolutionChamberRecipe(IFEU("laser_lens_sculk"),
-                new Ingredient.Value[]{
+        IFEURecipeBuilders.dissolutionChamberRecipe(ModContents.LASER_LENS_SCULK.get().getDefaultInstance())
+                .inputs(
                         tagValue(Tags.Items.GLASS_PANES_COLORLESS),
                         tagValue(Tags.Items.GLASS_PANES_COLORLESS),
                         tagValue(Tags.Items.GLASS_PANES_COLORLESS),
@@ -69,12 +71,14 @@ public class ModSerializableProvider extends TitaniumSerializableProvider {
                         itemValue(Items.SCULK.getDefaultInstance()),
                         itemValue(Items.SCULK.getDefaultInstance()),
                         itemValue(Items.SCULK.getDefaultInstance())
-                },new FluidStack(ModuleCore.ETHER.getSourceFluid().get(), 1000),200,
-                ModContents.LASER_LENS_SCULK.get().getDefaultInstance(),FluidStack.EMPTY
-        );
+                )
+                .inputFluid(new FluidStack(ModuleCore.ETHER.getSourceFluid().get(), 1000))
+                .processingTime(200)
+                .outputFluid(FluidStack.EMPTY)
+                .save();
 
-        new DissolutionChamberRecipe(IFEU("laser_lens_dragon"),
-                new Ingredient.Value[]{
+        IFEURecipeBuilders.dissolutionChamberRecipe(ModContents.LASER_LENS_DRAGON.get().getDefaultInstance())
+                .inputs(
                         tagValue(Tags.Items.GLASS_PANES_COLORLESS),
                         tagValue(Tags.Items.GLASS_PANES_COLORLESS),
                         tagValue(Tags.Items.GLASS_PANES_COLORLESS),
@@ -83,12 +87,14 @@ public class ModSerializableProvider extends TitaniumSerializableProvider {
                         itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
                         itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
                         itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance())
-                },new FluidStack(ModuleCore.ETHER.getSourceFluid().get(), 1000),200,
-                ModContents.LASER_LENS_DRAGON.get().getDefaultInstance(),FluidStack.EMPTY
-        );
+                )
+                .inputFluid(new FluidStack(ModuleCore.ETHER.getSourceFluid().get(), 1000))
+                .processingTime(200)
+                .outputFluid(FluidStack.EMPTY)
+                .save();
 
-        new DissolutionChamberRecipe(IFEU("ultimate_machine_frame"),
-                new Ingredient.Value[]{
+        IFEURecipeBuilders.dissolutionChamberRecipe(ModContents.ULTIMATE_MACHINE_FRAME.get().asItem().getDefaultInstance())
+                .inputs(
                         tagValue(IndustrialTags.Items.PLASTIC),
                         tagValue(IndustrialTags.Items.MACHINE_FRAME_SUPREME),
                         tagValue(IndustrialTags.Items.PLASTIC),
@@ -97,12 +103,14 @@ public class ModSerializableProvider extends TitaniumSerializableProvider {
                         itemValue(Items.DRAGON_EGG.getDefaultInstance()),
                         tagValue(ModTags.Items.GEARS_SCULK),
                         itemValue(Items.DRAGON_EGG.getDefaultInstance())
-                },new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),8000),200,
-                ModContents.ULTIMATE_MACHINE_FRAME.get().asItem().getDefaultInstance(),new FluidStack(Fluids.WATER,8000)
-        );
+                )
+                .inputFluid(new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),8000))
+                .processingTime(200)
+                .outputFluid(new FluidStack(Fluids.WATER,8000))
+                .save();
 
-        new DissolutionChamberRecipe(IFEU("empty_nether_star"),
-                new Ingredient.Value[]{
+        IFEURecipeBuilders.dissolutionChamberRecipe(new ItemStack(ModContents.EMPTY_NETHER_STAR.get(),4))
+                .inputs(
                         itemValue(Items.NETHER_STAR.getDefaultInstance()),
                         tagValue(IndustrialTags.Items.PLASTIC),
                         itemValue(Items.NETHER_STAR.getDefaultInstance()),
@@ -111,12 +119,14 @@ public class ModSerializableProvider extends TitaniumSerializableProvider {
                         itemValue(Items.NETHER_STAR.getDefaultInstance()),
                         tagValue(IndustrialTags.Items.PLASTIC),
                         itemValue(Items.NETHER_STAR.getDefaultInstance())
-                },new FluidStack(ModFluids.LIQUID_MALIC_ACID.getSourceFluid().get(), 4000), 200,
-                new ItemStack(ModContents.EMPTY_NETHER_STAR.get(),4),FluidStack.EMPTY
-        );
+                )
+                .inputFluid(new FluidStack(ModFluids.LIQUID_MALIC_ACID.getSourceFluid().get(), 4000))
+                .processingTime(200)
+                .outputFluid(FluidStack.EMPTY)
+                .save();
 
-        new DissolutionChamberRecipe(IFEU("rough_dragon_star"),
-                new Ingredient.Value[]{
+        IFEURecipeBuilders.dissolutionChamberRecipe(new ItemStack(ModContents.ROUGH_DRAGON_STAR.get(),2))
+                .inputs(
                         itemValue(ModContents.EMPTY_NETHER_STAR.get().getDefaultInstance()),
                         itemValue(Items.WITHER_SKELETON_SKULL.getDefaultInstance()),
                         itemValue(ModContents.EMPTY_NETHER_STAR.get().getDefaultInstance()),
@@ -124,13 +134,15 @@ public class ModSerializableProvider extends TitaniumSerializableProvider {
                         tagValue(Tags.Items.RODS_BLAZE),
                         tagValue(Tags.Items.STORAGE_BLOCKS_DIAMOND),
                         tagValue(Tags.Items.RODS_BLAZE),
-                        tagValue(Tags.Items.STORAGE_BLOCKS_DIAMOND),
-                },new FluidStack(ModFluids.DRAGON_STAR_ESSENCE.getSourceFluid().get(), 200), 200,
-                new ItemStack(ModContents.ROUGH_DRAGON_STAR.get(),2),FluidStack.EMPTY
-        );
+                        tagValue(Tags.Items.STORAGE_BLOCKS_DIAMOND)
+                )
+                .inputFluid(new FluidStack(ModFluids.DRAGON_STAR_ESSENCE.getSourceFluid().get(), 200))
+                .processingTime(200)
+                .outputFluid(FluidStack.EMPTY)
+                .save();
 
-        new DissolutionChamberRecipe(IFEU("liquid_malic_acid"),
-                new Ingredient.Value[]{
+        IFEURecipeBuilders.dissolutionChamberRecipe(new ItemStack(ModContents.APPLE_CORE.get(),8),"liquid_malic_acid")
+                .inputs(
                         itemValue(Items.APPLE.getDefaultInstance()),
                         itemValue(Items.APPLE.getDefaultInstance()),
                         itemValue(Items.APPLE.getDefaultInstance()),
@@ -139,248 +151,210 @@ public class ModSerializableProvider extends TitaniumSerializableProvider {
                         itemValue(Items.APPLE.getDefaultInstance()),
                         itemValue(Items.APPLE.getDefaultInstance()),
                         itemValue(Items.APPLE.getDefaultInstance())
-                },new FluidStack(ModuleCore.ETHER.getSourceFluid().get(), 100), 200,
-                new ItemStack(ModContents.APPLE_CORE.get(),8),new FluidStack(ModFluids.LIQUID_MALIC_ACID.getSourceFluid().get(), 10)
-        );
+                )
+                .inputFluid(new FluidStack(ModuleCore.ETHER.getSourceFluid().get(), 100))
+                .processingTime(200)
+                .outputFluid(new FluidStack(ModFluids.LIQUID_MALIC_ACID.getSourceFluid().get(), 10))
+                .save();
 
-        new DissolutionChamberRecipe(IFEU("nether_star"),
-                new Ingredient.Value[]{
+        IFEURecipeBuilders.dissolutionChamberRecipe(Items.NETHER_STAR.getDefaultInstance())
+                .inputs(
                         itemValue(ModContents.ROUGH_DRAGON_STAR.get().getDefaultInstance())
-                },new FluidStack(ModFluids.LIQUID_MALIC_ACID.getSourceFluid().get(), 1000), 200,
-                Items.NETHER_STAR.getDefaultInstance(),FluidStack.EMPTY
-        );
+                )
+                .inputFluid(new FluidStack(ModFluids.LIQUID_MALIC_ACID.getSourceFluid().get(), 1000))
+                .processingTime(200)
+                .outputFluid(FluidStack.EMPTY)
+                .save();
 
 
         //infuser
-        new InfuserRecipe(IFEU("dragon_star"),
-                Items.NETHER_STAR.getDefaultInstance(), new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(), 1000),200, ModContents.DRAGON_STAR.get().getDefaultInstance()
-        );
+        IFEURecipeBuilders.infuserRecipe(ModContents.DRAGON_STAR.get().getDefaultInstance())
+                .input(Items.NETHER_STAR.getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(), 1000))
+                .processingTime(200)
+                .save();
 
-        new InfuserRecipe(IFEU("dirt_mud"),
-                Items.DIRT.getDefaultInstance(),new FluidStack(Fluids.WATER.getSource(),1000),200,Items.MUD.getDefaultInstance()
-        );
 
-        new InfuserRecipe(IFEU("mud_clay"),
-                Items.MUD.getDefaultInstance(),new FluidStack(Fluids.WATER.getSource(),2000),500,Items.CLAY.getDefaultInstance()
-        );
+        IFEURecipeBuilders.infuserRecipe(Items.MUD.getDefaultInstance(),"dirt_mud")
+                .input(Items.DIRT.getDefaultInstance())
+                .inputFluid(new FluidStack(Fluids.WATER.getSource(),1000))
+                .processingTime(200)
+                .save();
 
-        new InfuserRecipe(IFEU("cobblestone_blackstone"),
-                Items.COBBLESTONE.getDefaultInstance(),new FluidStack(Fluids.LAVA.getSource(),200),200,Items.BLACKSTONE.getDefaultInstance()
-        );
+        IFEURecipeBuilders.infuserRecipe(Items.CLAY.getDefaultInstance(),"mud_clay")
+                .input(Items.MUD.getDefaultInstance())
+                .inputFluid(new FluidStack(Fluids.WATER.getSource(),2000))
+                .processingTime(500)
+                .save();
 
-        new InfuserRecipe(IFEU("sculk"),
-                Items.AMETHYST_SHARD.getDefaultInstance(),new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),200),200,Items.ECHO_SHARD.getDefaultInstance()
-        );
+        IFEURecipeBuilders.infuserRecipe(Items.BLACKSTONE.getDefaultInstance(),"cobblestone_blackstone")
+                .input(Items.COBBLESTONE.getDefaultInstance())
+                .inputFluid(new FluidStack(Fluids.LAVA.getSource(),200))
+                .processingTime(200)
+                .save();
 
-        new InfuserRecipe(IFEU("crying_obsidian"),
-                Items.OBSIDIAN.getDefaultInstance(),new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),2000),200,Items.CRYING_OBSIDIAN.getDefaultInstance()
-        );
+        IFEURecipeBuilders.infuserRecipe(Items.ECHO_SHARD.getDefaultInstance(),"sculk")
+                .input(Items.AMETHYST_SHARD.getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),200))
+                .processingTime(200)
+                .save();
+
+        IFEURecipeBuilders.infuserRecipe(Items.CRYING_OBSIDIAN.getDefaultInstance())
+                .input(Items.OBSIDIAN.getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),2000))
+                .processingTime(200)
+                .save();
+
+        IFEURecipeBuilders.infuserRecipe(Items.DRAGON_BREATH.getDefaultInstance())
+                .input(Items.GLASS_BOTTLE.getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),10))
+                .processingTime(200)
+                .save();
 
 
         //laser drill fluid
-        new LaserDrillFluidRecipe("liquid_sculk_matter",
-                new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),10).writeToNBT(new CompoundTag()),
-                Ingredient.of(ModContents.LASER_LENS_SCULK.get()),
-                MC("warden"),
-                new LaserDrillRarity(new ResourceKey[0], new ResourceKey[0],-64,256,8)
-        );
-        new LaserDrillFluidRecipe("liquid_dragon_breath",
-                new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),10).writeToNBT(new CompoundTag()),
-                Ingredient.of(ModContents.LASER_LENS_DRAGON.get()),
-                MC("ender_dragon"),
-                new LaserDrillRarity(LaserDrillRarity.END, new ResourceKey[0],-64,256,8)
-        );
+        IFEURecipeBuilders.laserDrillFluidRecipe(new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),10),"liquid_sculk_matter")
+                .catalyst(Ingredient.of(ModContents.LASER_LENS_SCULK.get()))
+                .entity(MC("warden"))
+                .rarity(new LaserDrillRarity(new ResourceKey[0], new ResourceKey[0],-64,256,8))
+                .save();
+        
+
+        IFEURecipeBuilders.laserDrillFluidRecipe(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),10),"liquid_dragon_breath")
+                .catalyst(Ingredient.of(ModContents.LASER_LENS_DRAGON.get()))
+                .entity(MC("ender_dragon"))
+                .rarity(new LaserDrillRarity(LaserDrillRarity.END, new ResourceKey[0],-64,256,8))
+                .save();
 
         //fluid extractor
-        new FluidExtractorRecipe(IFEU("dragon_star_essence"),
-                itemValue(new ItemStack(ModContents.DRAGON_STAR_BLOCK.get())),
-                Blocks.AIR,
-                0.01f,
-                new FluidStack(ModFluids.DRAGON_STAR_ESSENCE.getSourceFluid().get(),2),
-                false
-        );
+        IFEURecipeBuilders.fluidExtractorRecipe(new FluidStack(ModFluids.DRAGON_STAR_ESSENCE.getSourceFluid().get(),2),"dragon_star_essence")
+                .inputBlock(itemValue(new ItemStack(ModContents.DRAGON_STAR_BLOCK.get())))
+                .resultBlock(Blocks.AIR)
+                .breakChance(0.01f)
+                .isDefault(false)
+                .save();
 
         //arcane dragon egg forging
-        new ArcaneDragonEggForgingRecipe(IFEU("dead_dragon_egg"),
-                new ItemStack(Items.EGG,4),new FluidStack(ModuleCore.ESSENCE.getSourceFluid().get(),250),new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),4000),
-                200,ModContents.DEAD_DRAGON_EGG_ITEM.get().getDefaultInstance(),FluidStack.EMPTY
-        );
+        IFEURecipeBuilders.arcaneDragonEggForgingRecipe(ModContents.DEAD_DRAGON_EGG_ITEM.get().getDefaultInstance(),"dead_dragon_egg")
+                .input(new ItemStack(Items.EGG,4))
+                .inputFluids(new FluidStack(ModuleCore.ESSENCE.getSourceFluid().get(),250),new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),4000))
+                .outputFluid(FluidStack.EMPTY)
+                .processingTime(200)
+                .save();
 
-        new ArcaneDragonEggForgingRecipe(IFEU("egg_1"),
-                Items.DRAGON_EGG.getDefaultInstance(),new FluidStack(ModuleCore.PINK_SLIME.getSourceFluid().get(),100),new FluidStack(Fluids.WATER,1000),
-                200,new ItemStack(Items.EGG,16),new FluidStack(ModuleCore.ESSENCE.getSourceFluid().get(), 100)
-        );
+        IFEURecipeBuilders.arcaneDragonEggForgingRecipe(new ItemStack(Items.EGG,16),"egg_1")
+                .input(Items.DRAGON_EGG.getDefaultInstance())
+                .inputFluids(new FluidStack(ModuleCore.PINK_SLIME.getSourceFluid().get(),100),new FluidStack(Fluids.WATER,1000))
+                .outputFluid(new FluidStack(ModuleCore.ESSENCE.getSourceFluid().get(), 100))
+                .processingTime(200)
+                .save();
 
-        new ArcaneDragonEggForgingRecipe(IFEU("egg_2"),
-                ModContents.DEAD_DRAGON_EGG_ITEM.get().getDefaultInstance(),new FluidStack(ModuleCore.PINK_SLIME.getSourceFluid().get(),100),new FluidStack(Fluids.WATER,1000),
-                200,new ItemStack(Items.EGG,16),FluidStack.EMPTY
-        );
+        IFEURecipeBuilders.arcaneDragonEggForgingRecipe(new ItemStack(Items.EGG,16),"egg_2")
+                .input(ModContents.DEAD_DRAGON_EGG_ITEM.get().getDefaultInstance())
+                .inputFluids(new FluidStack(ModuleCore.PINK_SLIME.getSourceFluid().get(),100),new FluidStack(Fluids.WATER,1000))
+                .outputFluid(FluidStack.EMPTY)
+                .processingTime(200)
+                .save();
         
         //shaped
             //如果配方内某个格子为空,请使用ModContents.AIR.get().getDefaultInstance(),而不是ItemStack.EMPTY
-        new ShapedRecipe(IFEU("dragon_star_block"),
-                new Ingredient.Value[]{
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance())
-                },
-                new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),8000),
-                ModContents.DRAGON_STAR_BLOCK.get().asItem().getDefaultInstance()
-        );
-
-        new ShapedRecipe(IFEU("dragon_star_axe"),
-                new Ingredient.Value[]{
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(Items.NETHERITE_AXE.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(Items.STICK.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance())
-                },
-                new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000),
-                ModContents.DRAGON_STAR_AXE.get().getDefaultInstance()
-        );
-
-        new ShapedRecipe(IFEU("dragon_star_hoe"),
-                new Ingredient.Value[]{
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(Items.NETHERITE_HOE.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(Items.STICK.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance())
-                },
-                new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000),
-                ModContents.DRAGON_STAR_HOE.get().getDefaultInstance()
-        );
+        IFEURecipeBuilders.shapedRecipe(ModContents.DRAGON_STAR_BLOCK.get().asItem().getDefaultInstance())
+                .pattern("AAA")
+                .pattern("AAA")
+                .pattern("AAA")
+                .define('A', ModContents.DRAGON_STAR.get().getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),8000))
+                .save();
 
 
-        new ShapedRecipe(IFEU("dragon_star_pickaxe"),
-                new Ingredient.Value[]{
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(Items.NETHERITE_PICKAXE.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(Items.STICK.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance())
-                },
-                new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000),
-                ModContents.DRAGON_STAR_PICKAXE.get().getDefaultInstance()
-        );
-
-        new ShapedRecipe(IFEU("dragon_star_shovel"),
-                new Ingredient.Value[]{
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(Items.NETHERITE_SHOVEL.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(Items.STICK.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance())
-                },
-                new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000),
-                ModContents.DRAGON_STAR_SHOVEL.get().getDefaultInstance()
-        );
+        IFEURecipeBuilders.shapedRecipe(ModContents.DRAGON_STAR_AXE.get().getDefaultInstance())
+                .pattern("AA ")
+                .pattern("AB ")
+                .pattern(" C ")
+                .define('A', ModContents.DRAGON_STAR.get().getDefaultInstance())
+                .define('B', Items.NETHERITE_AXE.getDefaultInstance())
+                .define('C', Items.STICK.getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000))
+                .save();
 
 
-        new ShapedRecipe(IFEU("dragon_star_sword"),
-                new Ingredient.Value[]{
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(Items.NETHERITE_SWORD.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(Items.STICK.getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance())
-                },
-                new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000),
-                ModContents.DRAGON_STAR_SWORD.get().getDefaultInstance()
-        );
+        IFEURecipeBuilders.shapedRecipe(ModContents.DRAGON_STAR_HOE.get().getDefaultInstance())
+                .pattern("AA ")
+                .pattern(" B ")
+                .pattern(" C ")
+                .define('A', ModContents.DRAGON_STAR.get().getDefaultInstance())
+                .define('B', Items.NETHERITE_HOE.getDefaultInstance())
+                .define('C', Items.STICK.getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000))
+                .save();
 
-        new ShapedRecipe(IFEU("connect_tool"),
-                new Ingredient.Value[]{
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.DRAGON_STAR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        tagValue(Tags.Items.RODS),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        tagValue(Tags.Items.RODS),
-                        itemValue(ModContents.AIR.get().getDefaultInstance()),
-                        itemValue(ModContents.AIR.get().getDefaultInstance())
-                },
-                new FluidStack(ModuleCore.PINK_SLIME.getSourceFluid().get(),1000),
-                ModContents.CONNECT_TOOL.get().getDefaultInstance()
-        );
 
-        new ShapedRecipe(IFEU("big_dissolution_chamber_core"),
-                new Ingredient.Value[]{
-                        tagValue(IndustrialTags.Items.PLASTIC),
-                        itemValue(ModuleTransportStorage.BLACK_HOLE_TANK_SIMPLE.getLeft().get().asItem().getDefaultInstance()),
-                        tagValue(IndustrialTags.Items.PLASTIC),
-                        itemValue(ModuleCore.DISSOLUTION_CHAMBER.getLeft().get().asItem().getDefaultInstance()),
-                        tagValue(ModTags.Items.MACHINE_FRAME_ULTIMATE),
-                        itemValue(ModuleCore.DISSOLUTION_CHAMBER.getLeft().get().asItem().getDefaultInstance()),
-                        tagValue(IndustrialTags.Items.PLASTIC),
-                        itemValue(ModuleTransportStorage.BLACK_HOLE_UNIT_SIMPLE.getLeft().get().asItem().getDefaultInstance()),
-                        tagValue(IndustrialTags.Items.PLASTIC)
-                },
-                new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),1000),
-                ModBlocks.BIG_DISSOLUTION_CHAMBER_CORE.getLeft().get().asItem().getDefaultInstance()
-        );
+        IFEURecipeBuilders.shapedRecipe(ModContents.DRAGON_STAR_PICKAXE.get().getDefaultInstance())
+                .pattern("AAA")
+                .pattern(" B ")
+                .pattern(" C ")
+                .define('A', ModContents.DRAGON_STAR.get().getDefaultInstance())
+                .define('B', Items.NETHERITE_PICKAXE.getDefaultInstance())
+                .define('C', Items.STICK.getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000))
+                .save();
 
-        new ShapedRecipe(IFEU("black_hole_capacitor_advanced"),
-                new Ingredient.Value[]{
-                        tagValue(IndustrialTags.Items.PLASTIC),
-                        tagValue(IndustrialTags.Items.PLASTIC),
-                        tagValue(IndustrialTags.Items.PLASTIC),
-                        tagValue(Tags.Items.ENDER_PEARLS),
-                        itemValue(Items.ENDER_EYE.getDefaultInstance()),
-                        tagValue(Tags.Items.ENDER_PEARLS),
-                        tagValue(Tags.Items.STORAGE_BLOCKS_REDSTONE),
-                        tagValue(IndustrialTags.Items.MACHINE_FRAME_ADVANCED),
-                        tagValue(Tags.Items.STORAGE_BLOCKS_REDSTONE)
-                },
-                new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000),
-                ModBlocks.BLACK_HOLE_CAPACITOR_ADVANCED.getLeft().get().asItem().getDefaultInstance()
-        );
+        IFEURecipeBuilders.shapedRecipe(ModContents.DRAGON_STAR_SHOVEL.get().getDefaultInstance())
+                .pattern(" A ")
+                .pattern(" B ")
+                .pattern(" C ")
+                .define('A', ModContents.DRAGON_STAR.get().getDefaultInstance())
+                .define('B', Items.NETHERITE_SHOVEL.getDefaultInstance())
+                .define('C', Items.STICK.getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000))
+                .save();
 
-//        new ShapedRecipe(IFEU("black_hole_capacitor_supreme"),
-//                new Ingredient.Value[]{
-//                        tagValue(IndustrialTags.Items.PLASTIC),
-//                        tagValue(IndustrialTags.Items.PLASTIC),
-//                        tagValue(IndustrialTags.Items.PLASTIC),
-//                        tagValue(Tags.Items.ENDER_PEARLS),
-//                        itemValue(Items.ENDER_EYE.getDefaultInstance()),
-//                        tagValue(Tags.Items.ENDER_PEARLS),
-//                        tagValue(Tags.Items.STORAGE_BLOCKS_REDSTONE),
-//                        tagValue(IndustrialTags.Items.MACHINE_FRAME_SUPREME),
-//                        tagValue(Tags.Items.STORAGE_BLOCKS_REDSTONE)
-//                },
-//                new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),2000),
-//                ModBlocks.BLACK_HOLE_CAPACITOR_SUPREME.getLeft().get().asItem().getDefaultInstance()
-//        );
 
-        new ShapedRecipeBuilder(ModBlocks.BLACK_HOLE_CAPACITOR_SUPREME.getLeft().get().asItem().getDefaultInstance())
+        IFEURecipeBuilders.shapedRecipe(ModContents.DRAGON_STAR_SWORD.get().getDefaultInstance())
+                .pattern(" A ")
+                .pattern(" B ")
+                .pattern(" C ")
+                .define('A', ModContents.DRAGON_STAR.get().getDefaultInstance())
+                .define('B', Items.NETHERITE_SWORD.getDefaultInstance())
+                .define('C', Items.STICK.getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000))
+                .save();
+
+        IFEURecipeBuilders.shapedRecipe(ModContents.CONNECT_TOOL.get().getDefaultInstance())
+                .pattern("  A")
+                .pattern(" B ")
+                .pattern("B  ")
+                .define('A', ModContents.DRAGON_STAR.get().getDefaultInstance())
+                .define('B', Tags.Items.RODS)
+                .inputFluid(new FluidStack(ModuleCore.PINK_SLIME.getSourceFluid().get(),1000))
+                .save();
+
+        IFEURecipeBuilders.shapedRecipe(ModBlocks.BIG_DISSOLUTION_CHAMBER_CORE.getLeft().get().asItem().getDefaultInstance())
+                .pattern("ABA")
+                .pattern("CDC")
+                .pattern("AEA")
+                .define('A', IndustrialTags.Items.PLASTIC)
+                .define('B', ModuleTransportStorage.BLACK_HOLE_TANK_SIMPLE.getLeft().get().asItem().getDefaultInstance())
+                .define('C', ModuleCore.DISSOLUTION_CHAMBER.getLeft().get().asItem().getDefaultInstance())
+                .define('D', ModTags.Items.MACHINE_FRAME_ULTIMATE)
+                .define('E', ModuleTransportStorage.BLACK_HOLE_UNIT_SIMPLE.getLeft().get().asItem().getDefaultInstance())
+                .inputFluid(new FluidStack(ModFluids.LIQUID_SCULK_MATTER.getSourceFluid().get(),1000))
+                .save();
+
+
+        IFEURecipeBuilders.shapedRecipe(ModBlocks.BLACK_HOLE_CAPACITOR_ADVANCED.getLeft().get().asItem().getDefaultInstance())
+                .pattern("AAA")
+                .pattern("BCB")
+                .pattern("DED")
+                .define('A', IndustrialTags.Items.PLASTIC)
+                .define('B', Tags.Items.ENDER_PEARLS)
+                .define('C', Items.ENDER_EYE.getDefaultInstance())
+                .define('D', Tags.Items.STORAGE_BLOCKS_REDSTONE)
+                .define('E', IndustrialTags.Items.MACHINE_FRAME_ADVANCED)
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),1000))
+                .save();
+
+        IFEURecipeBuilders.shapedRecipe(ModBlocks.BLACK_HOLE_CAPACITOR_SUPREME.getLeft().get().asItem().getDefaultInstance())
                 .pattern("AAA")
                 .pattern("BCB")
                 .pattern("DED")
@@ -389,18 +363,14 @@ public class ModSerializableProvider extends TitaniumSerializableProvider {
                 .define('C', Items.ENDER_EYE.getDefaultInstance())
                 .define('D', Tags.Items.STORAGE_BLOCKS_REDSTONE)
                 .define('E', IndustrialTags.Items.MACHINE_FRAME_SUPREME)
-                .fluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),2000))
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),2000))
                 .save();
-
         //shapeless
             //有多少输入写多少
-        new ShapelessRecipe(IFEU("dragon_star_from_block"),
-                new Ingredient.Value[]{
-                        itemValue(ModContents.DRAGON_STAR_BLOCK.get().asItem().getDefaultInstance())
-                },
-                new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),100),
-                ModContents.DRAGON_STAR.get().getDefaultInstance().copyWithCount(9)
-        );
+        IFEURecipeBuilders.shapelessRecipe(ModContents.DRAGON_STAR.get().getDefaultInstance().copyWithCount(9),"dragon_star_from_block")
+                .inputs(itemValue(ModContents.DRAGON_STAR_BLOCK.get().asItem().getDefaultInstance()))
+                .inputFluid(new FluidStack(ModFluids.LIQUID_DRAGON_BREATH.getSourceFluid().get(),100))
+                .save();
 
         DissolutionChamberRecipe.RECIPES.forEach(dissolutionChamberRecipe -> map.put(dissolutionChamberRecipe,dissolutionChamberRecipe));
         InfuserRecipe.RECIPES.forEach(infuserRecipe -> map.put(infuserRecipe,infuserRecipe));