+++ changelog.md
@@ -126,6 +126,11 @@
 - 添加了新的插件(苹果插件,用于水培床增加苹果产量)
 - 添加了下界之星增产方法
 
+### v2.1.106
+- 为大部分配方添加了内建的builder类
+- 为苹果插件添加了KubeJS支持
+- 添加了灌注器内液态龙息+玻璃瓶合成龙息的配方
+
 ## en_us
 ### v1.0.0
 - init
@@ -250,3 +255,8 @@
 - Added some new recipes to the infuser
 - Added a new addon (Apple addon for increasing apple yield in hydroponic beds)
 - Added the Lower bound Star method for increasing production
+
+### v2.1.106
+- Added built-in builder class for most recipes
+- Added KubeJS support for Apple Addon
+- Added the recipe of liquid Dragon Breath in the infuser and Dragon Breath in glass bottles